requirements.txt
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
scikit-surprise==1.1.3
implicit==0.7.2
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
pyyaml==6.0.1
tqdm==4.65.0